[ALPS09685540] cam:FS: fix coverity issue
[Description]
Fix coverity issue

[Test]
build / local test pass

MTK-Commit-Id: 361508b9bc272fd8dcad01c166e1a1c2ec668890

CR-Id: ALPS09685540
Feature: [Android Default] Camera Application Basic Functions

+++ imgsensor/src-v4l2/frame-sync/frame_sync_algo.c
@@ -801,7 +801,7 @@ static unsigned int calc_vts_sync_bias_lc(const unsigned int idx)
  *     0: not trigger in ts critical section
  */
 static inline unsigned int check_timing_critical_section(
-	const unsigned int pred_vdiff, const unsigned int target_min_fl_us)
+	const long long pred_vdiff, const unsigned int target_min_fl_us)
 {
 	unsigned int threshold = 0/*, delta = 0*/;
